requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
aiosqlite
PyPDF2
langchain
faiss-cpu
openai
jinja2
pydantic-settings
langchain-community
greenlet
tiktoken